most kernels to polar_arrow
