+++ polars/polars-arrow/src/prelude.rs
@@ -1,2 +1,7 @@
 pub use crate::array::default_arrays::*;
 pub use crate::array::*;
+use arrow::array::{ListArray, Utf8Array};
+
+pub type AlignedVec<T> = arrow::buffer::MutableBuffer<T>;
+pub type LargeStringArray = Utf8Array<i64>;
+pub type LargeListArray = ListArray<i64>;
