+++ polars/polars-core/src/chunked_array/float.rs
@@ -1,6 +1,6 @@
-use crate::chunked_array::kernels::{is_finite, is_infinite, is_nan, is_not_nan};
 use crate::prelude::*;
 use num::Float;
+use polars_arrow::kernels::float::*;
 
 impl<T> ChunkedArray<T>
 where
